Create power_set_generator
The Power Set is the set that includes all possible subsets of a given set. As an example: The set [1, 2, 3] has 8 possible subsets: [1], [2], [3], [1, 2], [1, 3], [2, 3], [1, 2, 3], [[]]. It is a known mathematical fact that the number of elements in the power set is 2 raised to the power of the number of elements in the set. In the example above, there are 3 elements and the power set thus has 2^3 elements, as we see.

This code produces the power set of a given set.